Split thread helpers out
